c est long putain

+++ srcs/ft_strlen.c
@@ -1,6 +1,6 @@
 #include "../includes/libft.h"
 
-int		ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	char	*cpy;
